Made it so the DataDrivenPlugins will only load if their pluginid is -1 or they are allowed features.
git-svn-id: https://192.168.3.139:8443/svn/Medical/trunk@2413 956a8c8d-885e-cb46-a0b9-a2f355b4c364

+++ Standalone/Controller/OSHelper/LicenseManager.cs
@@ -139,7 +139,7 @@ namespace Medical
 #endif
         }
 
-        public bool allowFeature(int featureCode)
+        public bool allowFeature(long featureCode)
         {
 #if CRACKED
             return true;
